Police and Sprite
Police now aim and shoot at a zombie that is within vision. Police needs to go to call of screaming or if called by phone.

Added sprite for a phone booth for the humans to rush and call to.

Player can now click away from the target instead of locking onto the target until they die.

cleaned up all the code needs comments