User Updates
Camera updated to follow player on trigger. Camera off the follow trigger can move freely around the map.

Player movement is smooth. There is feedback on the movement and the targeting. The player is shown dying immediately. Feedback on attacking.

The RunawayScript is now HumanScript because I am combining the run to a booth with running to a chest to get weapons.

Made a particle system for the blood, it needs work.